+++ tests/FixturesInternalInterfaces/SecondStringable.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AsceticSoft\Wirebox\Tests\FixturesInternalInterfaces;
+
+class SecondStringable implements \Stringable
+{
+    public function __toString(): string
+    {
+        return 'second';
+    }
+}